requirements.txt
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
larksuite-oapi==1.0.34
openai==1.30.0
markdown==3.5.1
openpyxl==3.1.5
xmind==1.2.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.0.3
numpy==1.24.4
loguru==0.7.2
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
flask==2.3.3
Pillow==10.0.0 